+++ OpenOrchestrator/orchestrator/tabs/queue_tab.py
@@ -5,17 +5,19 @@ from nicegui import ui
 
 from OpenOrchestrator.database import db_util
 from OpenOrchestrator.database.queues import QueueStatus
+from OpenOrchestrator.orchestrator.datetime_input import DatetimeInput
 
 
-QUEUE_COLUMNS = (
+QUEUE_COLUMNS = [
     {'name': "Queue Name", 'label': "Queue Name", 'field': "Queue Name", 'align': 'left', 'sortable': True},
     {'name': "New", 'label': "New", 'field': "New", 'align': 'left', 'sortable': True},
     {'name': "In Progress", 'label': "In Progress", 'field': "In Progress", 'align': 'left', 'sortable': True},
     {'name': "Done", 'label': "Done", 'field': "Done", 'align': 'left', 'sortable': True},
-    {'name': "Failed", 'label': "Failed", 'field': "Failed", 'align': 'left', 'sortable': True}
-)
+    {'name': "Failed", 'label': "Failed", 'field': "Failed", 'align': 'left', 'sortable': True},
+    {'name': "Abandoned", 'label': "Abandoned", 'field': "Abandoned", 'align': 'left', 'sortable': True}
+]
 
-ELEMENT_COLUMNS = (
+ELEMENT_COLUMNS = [
     {'name': "Reference", 'label': "Reference", 'field': "Reference", 'align': 'left', 'sortable': True},
     {'name': "Status", 'label': "Status", 'field': "Status", 'align': 'left', 'sortable': True},
     {'name': "Data", 'label': "Data", 'field': "Data", 'align': 'left', 'sortable': True},
@@ -25,7 +27,7 @@ ELEMENT_COLUMNS = (
     {'name': "End Date", 'label': "End Date", 'field': "End Date", 'align': 'left', 'sortable': True},
     {'name': "Created By", 'label': "Created By", 'field': "Created By", 'align': 'left', 'sortable': True},
     {'name': "ID", 'label': "ID", 'field': "ID", 'align': 'left', 'sortable': True}
-)
+]
 
 
 # pylint: disable-next=too-few-public-methods
@@ -33,7 +35,7 @@ class QueueTab():
     """The 'Queues' tab object. It contains tables and buttons for dealing with queues."""
     def __init__(self, tab_name: str) -> None:
         with ui.tab_panel(tab_name):
-            self.queue_table = ui.table(title="Queues", columns=QUEUE_COLUMNS, rows=[], row_key='Queue Name', pagination=50).classes("w-full")
+            self.queue_table = ui.table(title="Queues", columns=QUEUE_COLUMNS, rows=[], row_key='Queue Name', pagination={'rowsPerPage': 50, 'sortBy': 'Queue Name'}).classes("w-full")
             self.queue_table.on("rowClick", self._row_click)
 
     def update(self):
@@ -48,7 +50,8 @@ class QueueTab():
                 "New": count.get(QueueStatus.NEW, 0),
                 "In Progress": count.get(QueueStatus.IN_PROGRESS, 0),
                 "Done": count.get(QueueStatus.DONE, 0),
-                "Failed": count.get(QueueStatus.FAILED, 0)
+                "Failed": count.get(QueueStatus.FAILED, 0),
+                "Abandoned": count.get(QueueStatus.ABANDONED, 0),
             }
             rows.append(row)
 
@@ -68,7 +71,8 @@ class QueuePopup():
 
         with ui.dialog(value=True).props('full-width full-height') as dialog, ui.card():
             with ui.row().classes("w-full"):
-                self._create_column_filter()
+                self.from_input = DatetimeInput("From Date", on_change=self._update, allow_empty=True).style('margin-left: 1rem')
+                self.to_input = DatetimeInput("To Date", on_change=self._update, allow_empty=True)
 
                 self.limit_select = ui.select(
                     options=[100, 200, 500, 1000, "All"],
@@ -76,9 +80,10 @@ class QueuePopup():
                     value=100,
                     on_change=self._update).classes("w-24")
 
-                ui.switch("Dense", on_change=lambda e: self._dense_table(e.value))
-
                 ui.space()
+
+                ui.switch("Dense", on_change=lambda e: self._dense_table(e.value))
+                self._create_column_filter()
                 ui.button(icon='refresh', on_click=self._update)
                 ui.button(icon="close", on_click=dialog.close)
             with ui.scroll_area().classes("h-full"):
@@ -111,6 +116,9 @@ class QueuePopup():
         if limit == 'All':
             limit = 1_000_000_000
 
-        queue_elements = db_util.get_queue_elements(self.queue_name, limit=limit)
+        from_date = self.from_input.get_datetime()
+        to_date = self.to_input.get_datetime()
+
+        queue_elements = db_util.get_queue_elements(self.queue_name, limit=limit, from_date=from_date, to_date=to_date)
         rows = [element.to_row_dict() for element in queue_elements]
         self.table.update_rows(rows)