+++ changelog.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-06-19
+
+### Added
+
+- 'Abandoned' status to queue elements.
+- 'Pausing' status added to triggers.
+- Disabling a 'Running' trigger in Orchestrator will set its status to 'Pausing'.
+- Scheduler will change a 'Pausing' trigger to 'Paused' when the process is done.
+- Queue elements can now be filtered by 'Created Date'.
+
+### Fixed
+
+- Refactoring for more correct typing.
+- Get-functions in db_util now works in SQLITE.
+- DatetimeInput tried to update before the UI was ready.
+- Credentials made with an old encryption key couldn't be edited.
+
+### Changed
+
+- Sort triggers & queues by name by default.
+- Log messages over 8000 characters are now truncated preserving the beginning and end.
+
 ## [1.2.0] - 2024-02-27
 
 ### Changed
@@ -73,7 +95,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/itk-dev-rpa/OpenOrchestrator/compare/1.2.0...HEAD
+[Unreleased]: https://github.com/itk-dev-rpa/OpenOrchestrator/compare/1.3.0...HEAD
+[1.3.0]: https://github.com/itk-dev-rpa/OpenOrchestrator/releases/tag/1.3.0
 [1.2.0]: https://github.com/itk-dev-rpa/OpenOrchestrator/releases/tag/1.2.0
 [1.1.0]: https://github.com/itk-dev-rpa/OpenOrchestrator/releases/tag/1.1.0
 [1.0.2]: https://github.com/itk-dev-rpa/OpenOrchestrator/releases/tag/1.0.2